refactor: remove unused components, move components
