+++ src/lib/app/settings/EmojiAvatarSelect.svelte
