+++ src/routes/(app)/me/settings/+page.svelte
@@ -6,9 +6,9 @@
 
   import { dispatch } from '$lib/dispatcher';
   import LL from '$src/i18n/i18n-svelte';
+  import EmojiAvatarSelect from '$src/lib/app/settings/EmojiAvatarSelect.svelte';
+  import SettingsEntry from '$src/lib/app/settings/SettingsEntry.svelte';
   import TopNavBar from '$src/lib/components/molecules/navigation/TopNavBar.svelte';
-  import EmojiAvatarSelect from '$src/lib/settings/EmojiAvatarSelect.svelte';
-  import SettingsEntry from '$src/lib/settings/SettingsEntry.svelte';
   import { state } from '$src/stores';
 
   import Envelope from '~icons/ph/envelope-fill';
